moved all account related pages to account and relatedAccount directories
