+++ src/app/modules/account/pages/search/component/partner-search/partner-search.component.scss
