+++ src/app/modules/account/pages/search/search.page.scss
