+++ src/app/modules/account/relatedAccount/pages/friends/friends.page.scss
