+++ src/app/modules/account/relatedAccount/pages/user-groups/user-groups.page.scss
